Round logic figured out
Added function playRound() which takes in the parameters of the human's
choice and the computer's choice, and does a bunch of checks to see who
wins in what scenario, writing line to the console to inform the human
player who won, also increments the score variables based on who wins.

Added variables to get the return of humanChoice and computerChoice to
then pass into playRound().