Final phase done, ready to push
playGame() function added, this completes the entire gameplay loop,
it simply calls the playRound() function 5 times, and resets the values
of humanSelection and computerSelection each time, keeps tabs on the
scores, and determines a winner at the end, and prints the final score.

playRound() function modified so it also prints the
overall score at each given round.